refactor: 리뷰 목록 조회와 리뷰 요약 조회 API 분리

+++ src/main/java/org/example/sejonglifebe/review/Review.java
@@ -25,7 +25,7 @@ public class Review {
     @Column(name = "review_id")
     private Long id;
 
-    private Double rating;
+    private Long rating;
 
     @Column(nullable = false)
     private String content;
@@ -51,13 +51,13 @@ public class Review {
     List<ReviewTag> reviewTags = new ArrayList<>();
 
     @Builder
-    private Review(Place place, Double rating, String content) {
+    private Review(Place place, Long rating, String content) {
         this.place = place;
         this.rating = rating;
         this.content = content;
     }
 
-    public static Review createReview(Place place, Double rating, String content, List<Tag> tags, List<String> imageUrls) {
+    public static Review createReview(Place place, Long rating, String content, List<Tag> tags, List<String> imageUrls) {
 
         Review review = Review.builder()
                 .place(place)
