+++ src/main/java/org/example/sejonglifebe/review/dto/RatingCount.java
@@ -1,6 +1,7 @@
 package org.example.sejonglifebe.review.dto;
 
 public record RatingCount(
-        Double rating,
+        Long rating,
         Long count
-) {}
+) {
+}
